Current Version of the IK-Chain

+++ rts/Sim/Units/Unit.h
@@ -93,7 +93,7 @@ public:
 	
 	float ikIDPool = 0;
 	IkChain * getIKChain(float ikID);
-	float CreateIKChain(LocalModelPiece* startPiece, float startPieceID, float endPieceID);
+	float CreateIKChain(LocalModelPiece* startPiece, unsigned int startPieceID, unsigned int  endPieceID);
 	void SetIKActive(float ikID, bool Active);
 	void SetIKGoal(float ikID, float goalX, float goalY, float goalZ);
 	void SetIKPieceSpeed(float ikID, float pieceID, float velX, float velY, float velZ);
